Refactor : Rename 'key_validation.py' to 'custom_exception.py'
