+++ internal/admins.py
@@ -1,7 +1,7 @@
 import datetime
 from fastapi import APIRouter, status, HTTPException, Depends, Request
 from database import get_db
-from internal.key_validation import ItemKeyValidationError, ForeignKeyValidationError
+from internal.custom_exception import ItemKeyValidationError, ForeignKeyValidationError
 from internal.schema import *
 from internal.crudf import *
 from models import Admin, Book, BookInfo, BookRequest, User, Notice, Category
