+++ internal/custom_exception.py
