+++ internal/schema.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel, validator
 from typing import List
-from internal.key_validation import InvalidDateFormatError
+from internal.custom_exception import InvalidDateFormatError
 import datetime
 
 # get_begin, get_end QUERY class
